+++ app/i18n/zh-cn/sub.php
@@ -1,10 +1,14 @@
 <?php
 
 return array(
+	'api' => array(
+		'documentation' => '复制以下地址，可供外部工具使用',
+		'title' => 'API',
+	),
 	'bookmarklet' => array(
-		'documentation' => 'Drag this button to your bookmarks toolbar or right-click it and choose "Bookmark This Link". Then click "Subscribe" button in any page you want to subscribe to.',// TODO
-		'label' => 'Subscribe',// TODO
-		'title' => 'Bookmarklet',// TODO
+		'documentation' => '拖动此书签到你的书签栏或者右键选择“收藏此链接”，然后在你想要订阅的页面上点击“订阅”按钮',
+		'label' => '订阅',
+		'title' => '书签应用',
 	),
 	'category' => array(
 		'_' => '分类',
@@ -18,20 +22,20 @@ return array(
 		'archiving' => '存档',
 		'auth' => array(
 			'configuration' => '认证',
-			'help' => '连接启用 HTTP 认证的 RSS 源',
+			'help' => '用于连接启用 HTTP 认证的 RSS 源',
 			'http' => 'HTTP 认证',
 			'password' => 'HTTP 密码',
 			'username' => 'HTTP 用户名',
 		),
-		'css_help' => '获取全文（注意，会耗费更多时间！）',
-		'css_path' => '原网站中文章的 CSS 路径',
+		'css_help' => '用于获取全文（注意，这将耗费更多时间！）',
+		'css_path' => '原文的 CSS 选择器',
 		'description' => '描述',
 		'empty' => '此源为空。请确认它是否正常更新。',
-		'error' => '此源遇到一些问题。请确认它是否可访问后重试。',
+		'error' => '此源遇到一些问题。请在确认是否能正常访问后重试。',
 		'in_main_stream' => '在首页中显示',
 		'informations' => '信息',
 		'keep_history' => '至少保存的文章数',
-		'moved_category_deleted' => '删除分类时，其中的 RSS 源会自动归类到<em>%s</em>。',
+		'moved_category_deleted' => '删除分类时，其中的 RSS 源会自动归类到 <em>%s</em>',
 		'no_selected' => '未选择 RSS 源。',
 		'number_entries' => '%d 篇文章',
 		'stats' => '统计',
@@ -45,8 +49,8 @@ return array(
 		'pubsubhubbub' => 'PubSubHubbub 即时通知',
 	),
 	'firefox' => array(
-		'documentation' => 'Follow the steps described <a href="https://developer.mozilla.org/en-US/Firefox/Releases/2/Adding_feed_readers_to_Firefox#Adding_a_new_feed_reader_manually">here</a> to add FreshRSS to Firefox feed reader list.',// TODO
-		'title' => 'Firefox feed reader',// TODO
+		'documentation' => '按照 <a href="https://developer.mozilla.org/en-US/Firefox/Releases/2/Adding_feed_readers_to_Firefox#Adding_a_new_feed_reader_manually">这里</a> 描述的步骤可将 FreshRSS 添加到 Firefox 阅读器列表',
+		'title' => 'Firefox RSS 阅读器',
 	),
 	'import_export' => array(
 		'export' => '导出',
@@ -63,11 +67,11 @@ return array(
 		'bookmark' => '订阅 (FreshRSS 书签)',
 		'import_export' => '导入/导出',
 		'subscription_management' => '订阅管理',
-		'subscription_tools' => 'Subscription tools',// TODO
+		'subscription_tools' => '订阅工具',
 	),
 	'title' => array(
 		'_' => '订阅管理',
 		'feed_management' => 'RSS 源管理',
-		'subscription_tools' => 'Subscription tools',// TODO
+		'subscription_tools' => '订阅工具',
 	),
 );