+++ lib/Minz/Dispatcher.php
@@ -10,7 +10,6 @@
  * C'est un singleton
  */
 class Minz_Dispatcher {
-	const CONTROLLERS_PATH_NAME = '/Controllers';
 
 	/* singleton */
 	private static $instance = null;
@@ -149,7 +148,7 @@ class Minz_Dispatcher {
 	 */
 	private static function loadController($base_name) {
 		$base_path = self::$registrations[$base_name];
-		$controller_filename = $base_path . '/controllers/' . $base_name . 'Controller.php';
+		$controller_filename = $base_path . '/Controllers/' . $base_name . 'Controller.php';
 		include_once $controller_filename;
 	}